+++ tests/Samples/HasOptionsEnum.php
@@ -3,11 +3,13 @@
 namespace Vixen\Enums\Tests\Samples;
 
 use Vixen\Enums\Attributes\Label;
+use Vixen\Enums\HasDescriptions;
 use Vixen\Enums\HasLabels;
 use Vixen\Enums\HasOptions;
 
 enum HasOptionsEnum: string
 {
+    use HasDescriptions;
     use HasLabels;
     use HasOptions;