+++ CHANGELOG.md
@@ -1,8 +1,21 @@
 # Changelog
 All notable changes to `laravel-enums` will be documented in this file.
 
-### 1.0.1 — 2025-05-08
-- Add support for Laravel 12
+### 1.1.0 • 2026-03-29
+- Renamed `Only` to `Filterable`
+- Added `HasValues` trait
+- Added `HasDescriptions` trait
+- Added `except` to the `Filterable` trait
+- Labels are now cached
+- All methods that return an array now return a collection
+- Fixed various minor bugs and optimized code
+- Removed the service provider
 
-### 1.0.0 — 2025-01-13
+### 1.0.2 • 2026-03-20
+- Added support for Laravel 13
+
+### 1.0.1 • 2025-05-08
+- Added support for Laravel 12
+
+### 1.0.0 • 2025-01-13
 Initial release.